% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{add_peak}
\alias{add_peak}
\title{Add peak.}
\usage{
add_peak(signal, peak, start)
}
\arguments{
\item{signal}{The original signal.}

\item{peak}{A vector containing the peak that wants o be added.}

\item{start}{The position (index) in the signal where the peak will begin.}
}
\value{
The signal with the added peak.
}
\description{
Adds a peak to a signal in the desired place.
}
\examples{
signal <- rep(0, 150)
peak <- generate_peak()
add_peak(signal = signal, peak = peak, start = 25)
}
