% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-package.R
\docType{package}
\name{align-package}
\alias{align}
\alias{align-package}
\title{align: Grid Specification Primitive}
\description{
A single core object for grid specification defined by dimension and extent. Resolution, coordinates, cell indices, and row/column mappings are derived on demand by arithmetic. Crop, expand, resize, and tile operations produce new specs on the same lattice. Pure index arithmetic, no data, no I/O.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
