% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-package.R
\docType{package}
\name{align}
\alias{align}
\alias{align-package}
\title{Data Alignment Tools}
\description{
Methods for the alignment and merging of partially or inexactly
aligned data.
}
\details{
This is a splat packed in-development package.

Package functions:

Two object (x, y) alignment: nAlign, cAlign, cowAlign.
One object (x) modification: regularize, warp.
}
