% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{alignment}
\alias{alignment}
\alias{plot.alignment}
\alias{print.alignment}
\alias{summary.alignment}
\title{alignment}
\usage{
\method{plot}{alignment}(x, ...)

\method{print}{alignment}(x, ...)

\method{summary}{alignment}(object, ...)
}
\arguments{
\item{x, object}{\code{alignment} class object}

\item{...}{Other arguments}
}
\description{
Generic methods for use with \code{alignment} class
objects generated with \code{..._align(...,  output = "alignment")}.
}
