% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{apply_pow}
\alias{apply_pow}
\title{Apply POW}
\usage{
apply_pow(X, y, lambdas, max_it = 1000)
}
\arguments{
\item{X}{A matrix with the samples to align}

\item{y}{The reference signal}

\item{lambdas}{the lambda to be used for the alignment of each sample}

\item{max_it}{maximum iterations to be done in the POW alignment}
}
\value{
a list with:
warped_samples
warps
}
\description{
Apply POW
}
