% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R, R/generics.R
\name{as_vast.vast}
\alias{as_vast.vast}
\alias{as_vast.SpatRaster}
\alias{as_vast.BasicRaster}
\alias{as_vast.list}
\alias{as_vast}
\title{Coerce to a vast grid specification}
\usage{
\method{as_vast}{vast}(x, ...)

\method{as_vast}{SpatRaster}(x, ...)

\method{as_vast}{BasicRaster}(x, ...)

\method{as_vast}{list}(x, ...)

as_vast(x, ...)
}
\arguments{
\item{x}{object to coerce}

\item{...}{passed to methods}
}
\value{
a \code{vast} object
}
\description{
Coerce to a vast grid specification
}
