% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candidate_data}
\alias{candidate_data}
\title{Synthetic Candidate Data}
\format{
## `candidate_data`
An array with 50 rows and 2 columns:
\describe{
  \item{d13c}{Synthetic d13C values}
  \item{m}{Synthetic meterage values}
  ...
}
}
\source{
Cedric Hagen
}
\usage{
candidate_data
}
\description{
A synthetic record, resembling d13C data, for testing the DTW algorithm
}
\keyword{datasets}
