% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{cell_from_rowcol}
\alias{cell_from_rowcol}
\title{Convert row and column to cell index}
\usage{
cell_from_rowcol(g, row, col)
}
\arguments{
\item{g}{a \code{vast} object}

\item{row}{integer row indices (1-based, from top)}

\item{col}{integer column indices (1-based, from left)}
}
\value{
integer vector of cell indices
}
\description{
Convert row and column to cell index
}
