% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{col_from_x}
\alias{col_from_x}
\title{Column index from x coordinate}
\usage{
col_from_x(g, x)
}
\arguments{
\item{g}{a \code{vast} object}

\item{x}{numeric vector of x coordinates}
}
\value{
integer column indices (1-based), \code{NA} if outside extent
}
\description{
Column index from x coordinate
}
