% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{compute_val_error}
\alias{compute_val_error}
\title{Compute validation error.}
\usage{
compute_val_error(
  X,
  y,
  W,
  iv,
  lambdas,
  fom = "rms",
  lambda1 = 10^6,
  parallel = TRUE
)
}
\arguments{
\item{X}{Matrix with data to be aligned (one sample per row).}

\item{y}{Reference sample.}

\item{W}{A matrix with of weights.}

\item{iv}{indexes of the points used to validate.}

\item{lambdas}{a vector with the penalties to test the POW.}

\item{fom}{figure of merit to use ('rms' or 'cor').}

\item{lambda1}{Penalty to avoid time inversions.}

\item{parallel}{Enable parallelization with BiocParallel}
}
\value{
a list with 2 values 'e_ix' and 'ti_ix'.
}
\description{
Computes validation error and checks for time inversion in warp function.
}
