% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_align.R
\name{cor_align}
\alias{cor_align}
\alias{cor_align.default}
\title{cor_align}
\usage{
\method{cor_align}{default}(x, y = NULL, by = NULL, min.overlap = NULL, ...)
}
\arguments{
\item{x}{First \code{vector} or \code{data.frame}, to be aligned
and merged with.}

\item{y}{Second \code{vector} or \code{data.frame}, to align and
merge with \code{x}.}

\item{by}{If \code{x} or \code{y} are \code{data.frame}s, the names
of the columns that \code{cor_align} should align with.}

\item{min.overlap}{By default, \code{cor_align} compares
correlations for lags with \code{x}/\code{y} overlaps of
2000 rows or 20\% if smaller. Other minimum (row number)
overlaps can be set using the extra argument min.overlap.}

\item{...}{Other arguments, typically handled by common
\code{alignment} functions/methods or ignored.}
}
\value{
By default, \code{cor_align} returns \code{x} and \code{y}
as an aligned \code{data.frame}. It also provides a lag correlation
profile (plot) and alignment report. The extra \code{function},
\code{output}, can be used to modify this behavior.
}
\description{
Correlation based (linear) alignment, row
offsetting second of two \code{vector}s or \code{data.frame}s
based on best fit as estimated by correlation coefficient.
}
\note{
\code{cor_align} is based on \code{cAlign} function in
earlier version of \code{pems.utils}. It works with vectors,
data.frames and other objects that can be converted to data.frame
with as.data.frame(), and uses \code{Rcpp} to run \code{C++} code.
}
\author{
Karl Ropkins
}
