% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{cow}
\alias{cow}
\title{cow}
\usage{
cow(Ta, X, Seg, Slack, Options)
}
\arguments{
\item{Ta}{First \code{vector}, target for alignment.}

\item{X}{Second \code{vector} to be aligned with target.}

\item{Seg}{Segment size for warping.}

\item{Slack}{Segment size expansion/compression range.}

\item{Options}{Other options (to documet).}
}
\description{
Non-linear alignment using Correlation Optimized
warping.
}
\note{
\code{cow} is based on similar matlab (to document)
}
\references{
(to document)
}
\author{
Daniel Quiroz Moreno
}
