% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{create_synthetic_dataset}
\alias{create_synthetic_dataset}
\title{Create Synthetic dataset.}
\usage{
create_synthetic_dataset(
  n_samples,
  n_peaks,
  length_out,
  mov_peaks,
  intensity,
  random_intensity
)
}
\arguments{
\item{n_samples}{a}

\item{n_peaks}{a}

\item{length_out}{a}

\item{mov_peaks}{a}

\item{intensity}{a}

\item{random_intensity}{a}
}
\value{
synthetic dataset.
}
\description{
Create Synthetic dataset.
}
