% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/operations.R
\name{crop}
\alias{crop}
\alias{crop.vast}
\title{Crop a grid to a smaller extent}
\usage{
crop(x, ...)

\method{crop}{vast}(x, extent, ...)
}
\arguments{
\item{x}{a \code{vast} object}

\item{...}{passed to methods}

\item{extent}{numeric \code{c(xmin, xmax, ymin, ymax)} to crop to}
}
\value{
a \code{vast} object
}
\description{
Restrict the extent, snapping inward to cell boundaries.
Resolution is unchanged.
}
