% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{draw_compare}
\alias{draw_compare}
\title{Draw the plot of the two mathed string}
\usage{
draw_compare(x, y)
}
\arguments{
\item{x}{A string}

\item{y}{A string}
}
\value{
ggplot of the matched string
}
\description{
Draw the plot of the two mathed string
}
\examples{
draw_compare("AG-T", "AGCT")
draw_compare("GAGCGT", "GA-C-T")
}
