% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw_r.R
\name{dtw_r}
\alias{dtw_r}
\title{dtw_r}
\usage{
dtw_r(target, candidate, g, edge)
}
\arguments{
\item{target}{The target time series}

\item{candidate}{The candidate time series}

\item{g}{The g parameter value for alignment}

\item{edge}{The edge parameter value for alignment}
}
\value{
The aligned candidate dataset (ri,t_out) and the corresponding Pearson's correlation coefficient (xc).
}
\description{
dtw_r
}
\examples{
dtw_r(target_data,candidate_data,0.98,0.15)
}
