% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_align.R
\name{fit_align}
\alias{fit_align}
\alias{fit_align.default}
\title{fit_align}
\usage{
fit_align(x, y = NULL, by = NULL, fun, lower, upper, ...)
}
\arguments{
\item{x}{A first time-series \code{vector} or a \code{data.frame}
containing time-series to use a reference when warp-fitting
\code{y}.}

\item{y}{A second time-series \code{vector} or a \code{data.frame}
containing time-series to warp-fit with \code{x}.}

\item{by}{If \code{x} or \code{y} are \code{data.frames}, the names
of time-series to warp-fit.}

\item{fun}{A \code{function} describing the warp-fitting model
to be applied.}

\item{lower}{The lowest values for \code{fun}, one per parameter.}

\item{upper}{The highest values for \code{fun}, one per parameter.}

\item{...}{Other arguments currently include:
\describe{
  \item{\code{output}}{The default \code{..._align} \code{output}
  is \code{c("summary", "plot", "ans")}. \code{output} options
   include: \code{"ans"}, \code{"plot"}, \code{"summary"} and
   \code{"alignment"}. Multiple \code{output}s are allowed, but
   only the last is captured by return. \code{alignment} is a
   special object class which may be helpful to those looking at
   alignments in more detail.}
}}
}
\description{
Time warp a data-series using a time-offsetting model.
}
\author{
Karl Ropkins
}
