% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{generate_peak}
\alias{generate_peak}
\title{Peak generator.}
\usage{
generate_peak(intensity = 1, length_peak = 100, alpha = 0, omega = 1)
}
\arguments{
\item{intensity}{Maximum intensity of the peak.}

\item{length_peak}{Length of the peak in indexes.}

\item{alpha}{for skewed normal distribution, modifies the first half of the
peak}

\item{omega}{for skewed normal distribution, modifies with of the peak}
}
\value{
A vector with a peak.
}
\description{
Generates a peak with normal distribution.
}
\keyword{internal}
