% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{getColor}
\alias{getColor}
\title{Get color of the character}
\usage{
getColor(c)
}
\arguments{
\item{c}{A character}
}
\value{
the color string of the character
}
\description{
Get color of the character
}
\examples{
getColor("A")
getColor("G")
getColor("C")
}
