% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation}
\alias{interpolation}
\title{Linear interpolation.}
\usage{
interpolation(t, y)
}
\arguments{
\item{t}{points at which to compute interpolated values.}

\item{y}{signal to be interpolated.}
}
\value{
a list with:
f: the interpolated signal,
s: points  of t with  1<= t & t <= length of y
g: gradient of y at points t.
}
\description{
Interpolates a vector.
}
\examples{
t <- c(1:10)
y <- c(1:10)
interpolation(t,y)
}
