% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast.R
\name{is_vast}
\alias{is_vast}
\title{Test if an object is a vast grid specification}
\usage{
is_vast(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical
}
\description{
Test if an object is a vast grid specification
}
