% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_align.R
\name{n_align}
\alias{n_align}
\title{n_align}
\usage{
n_align(x, y = NULL, n = 0, by = NULL, ...)
}
\arguments{
\item{x}{First \code{vector} or \code{data.frame}, to be aligned
and merged with.}

\item{y}{Second \code{vector} or \code{data.frame} to align and
merge with \code{x}.}

\item{n}{\code{y} offset when binding \code{y} to \code{x},
default 0.}

\item{by}{If \code{x} or \code{y} are \code{data.frame}s, the names
of the columns that \code{n_align} should align.}

\item{...}{Other arguments, currently ignored.}
}
\value{
\code{data.frame} of \code{x} and \code{y}, with \code{y}
offset \code{n} rows.
}
\description{
Basic alignment, row offsetting second of two
\code{vector}s or \code{data.frame}s.
}
\author{
Karl Ropkins
}
