% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{offset}
\alias{offset}
\title{Integer offset between aligned grids}
\usage{
offset(child, parent, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{child, parent}{\code{vast} objects with the same resolution}

\item{tol}{numeric tolerance}
}
\value{
named integer vector \code{c(col = ..., row = ...)}
}
\description{
Returns the column and row offset from the origin of \code{parent} to the
origin of \code{child}. Both grids must have the same resolution.
}
