% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimize_params_pow}
\alias{optimize_params_pow}
\title{Optimize parameters POW}
\usage{
optimize_params_pow(n_samples, params, ti_ix, e_ix)
}
\arguments{
\item{n_samples}{number of samples or signals to be used}

\item{params}{parametros}

\item{ti_ix}{matrix with time inversions}

\item{e_ix}{error matrix}
}
\value{
a list with two values best_params and params_ix
}
\description{
Optimize parameters POW
}
