% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize.R
\name{regularize}
\alias{regularize}
\title{regularize}
\usage{
regularize(x, ts = NULL, Hz = 1, ...)
}
\arguments{
\item{x}{the supplied \code{vector} or \code{data.frame}.}

\item{ts}{vector of sample times, one per row of \code{x}.}

\item{Hz}{The (regular) resolution of the output, default \code{1}
generates one per second.}

\item{...}{Other arguments, currently ignored.}
}
\description{
regularizes a supplied \code{vector} or \code{data.frame}
on basis on an supplied counter, timestamp, etc.
}
\note{
If \code{counter} is numeric, it is assumed to be reported in
1 second units.
}
\author{
Karl Ropkins
}
