% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{rowcol_from_cell}
\alias{rowcol_from_cell}
\title{Convert cell index to row and column}
\usage{
rowcol_from_cell(g, cell)
}
\arguments{
\item{g}{a \code{vast} object}

\item{cell}{integer vector of cell indices (1-based)}
}
\value{
two-column integer matrix: \code{row}, \code{col}
}
\description{
Convert cell index to row and column
}
