% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{rowcol_from_xy}
\alias{rowcol_from_xy}
\title{Convert xy coordinates to row and column}
\usage{
rowcol_from_xy(g, xy)
}
\arguments{
\item{g}{a \code{vast} object}

\item{xy}{two-column numeric matrix}
}
\value{
two-column integer matrix: \code{row}, \code{col}
}
\description{
Points outside the extent return \code{NA}.
}
