% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_reference.R
\name{select_reference}
\alias{select_reference}
\title{Select reference sample}
\usage{
select_reference(data, figure = "correlation")
}
\arguments{
\item{data}{a matrix with all the samples (by row)}

\item{figure}{'variance' or 'correlation' figure of merit to select best reference sample}
}
\value{
index of best reference sample
}
\description{
Select reference sample
}
