% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R, R/generics.R
\name{snap.vast}
\alias{snap.vast}
\alias{snap}
\title{Snap a grid to a parent lattice}
\usage{
\method{snap}{vast}(x, parent, ...)

snap(x, ...)
}
\arguments{
\item{x}{a \code{vast} object}

\item{parent}{\code{vast} object whose lattice to snap to}

\item{...}{passed to methods}
}
\value{
a \code{vast} object
}
\description{
Snap a grid to a parent lattice
}
