% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_signal.R
\name{synthetic_signal}
\alias{synthetic_signal}
\title{Synthetic signal.}
\usage{
synthetic_signal(
  peaks = 1,
  length_out = 1000,
  intensity = 1000,
  width = 5,
  length_peaks = NULL,
  baseline = NULL,
  noise = NULL
)
}
\arguments{
\item{peaks}{Can be an scalar with the number of peaks wanted in the signal
or a vector with the starting points for each peak.}

\item{length_out}{The final length of the desired signal (in indexes).}

\item{intensity}{The maximum intensity of the peaks, if is a scalar all
peaks will have the same intensity, if a vector each value will correspond to
a peak.}

\item{width}{The width of the peaks, if a scalar the same width will be used
for all peaks, if a vector each value will correspond to a peak. The parameter
will be passed to function 'dnorm()' where sd = length_peak / width.}

\item{length_peaks}{The length of each peak (in indexes), if a scalar all
peaks will have the same length, if a vector each value will correspond to
a peak.}

\item{baseline}{A vector that will be added to the base signal, if != NULL
length_out will be ignored and calculated with the length of this vector.}

\item{noise}{An scalar with the intensity of the noise desired.}
}
\value{
A vector with the designed signal.
}
\description{
Generates a synthetic signal with peaks with normal distribution.
}
\examples{
synthetic_signal()
synthetic_signal(peaks = 5)
synthetic_signal(intensity = c(100, 600, 1000), baseline = seq(0, 150, by = 0.1))
synthetic_signal(peaks = 10, length_out = 1500, noise = 10)
}
