% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_data}
\alias{target_data}
\title{Synthetic Target Data}
\format{
## `target_data`
An array with 251 rows and 2 columns:
\describe{
  \item{d13c}{Synthetic d13C values}
  \item{m}{Synthetic meterage values}
  ...
}
}
\source{
Cedric Hagen
}
\usage{
target_data
}
\description{
A synthetic record, resembling d13C data, for testing the DTW algorithm
}
\keyword{datasets}
