% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/tiles.R
\name{tiles}
\alias{tiles}
\alias{tiles.vast}
\title{Partition a grid into tiles}
\usage{
tiles(x, ...)

\method{tiles}{vast}(x, tile_dim = c(256L, 256L), ...)
}
\arguments{
\item{x}{a \code{vast} object}

\item{...}{passed to methods}

\item{tile_dim}{integer \code{c(ncol, nrow)} for each tile. Edge tiles
may be smaller.}
}
\value{
list of \code{vast} objects
}
\description{
Partition a grid into tiles
}
