% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{val}
\alias{val}
\title{Validation}
\usage{
val(X, y, W, iv, lambdas, fom = "rms", lambda1 = 10^6)
}
\arguments{
\item{X}{Matrix with data to be aligned (one sample per row).}

\item{y}{Reference sample.}

\item{W}{A matrix with of weights.}

\item{iv}{indexes of the points used to validate.}

\item{lambdas}{a vector with the penalties to test the POW.}

\item{fom}{figure of merit to use ('rms' or 'cor').}

\item{lambda1}{Penalty to avoid time inversions.}
}
\value{
list with e and i
}
\description{
Validation
}
\keyword{internal}
