% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{vast_res}
\alias{vast_res}
\title{Create a vast from extent and resolution}
\usage{
vast_res(extent, resolution)
}
\arguments{
\item{extent}{numeric \code{c(xmin, xmax, ymin, ymax)}}

\item{resolution}{numeric resolution, scalar (square cells) or
\code{c(dx, dy)}}
}
\value{
a \code{vast} object
}
\description{
Dimension is derived. Extent is adjusted (expanded) to honour exact
resolution — the returned extent may be larger than the input if the
extent is not evenly divisible by the resolution.
}
\examples{
vast_res(c(140, 160, -50, -30), 0.04)
}
