% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R, R/generics.R
\name{x.vast}
\alias{x.vast}
\alias{x}
\title{Cell-centre x coordinates}
\usage{
\method{x}{vast}(x, ...)

x(x, ...)
}
\arguments{
\item{x}{a \code{vast} object}

\item{...}{ignored}
}
\value{
numeric vector of length \code{ncol(x)}
}
\description{
Cell-centre x coordinates
}
