% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{x_from_col}
\alias{x_from_col}
\title{X coordinate from column index (cell centre)}
\usage{
x_from_col(g, col)
}
\arguments{
\item{g}{a \code{vast} object}

\item{col}{integer column indices (1-based)}
}
\value{
numeric x coordinates
}
\description{
X coordinate from column index (cell centre)
}
