% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{xy_from_cell}
\alias{xy_from_cell}
\title{Convert cell indices to xy coordinates (cell centres)}
\usage{
xy_from_cell(g, cell)
}
\arguments{
\item{g}{a \code{vast} object}

\item{cell}{integer vector of cell indices (1-based)}
}
\value{
two-column numeric matrix
}
\description{
Convert cell indices to xy coordinates (cell centres)
}
