% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R, R/generics.R
\name{y.vast}
\alias{y.vast}
\alias{y}
\title{Cell-centre y coordinates}
\usage{
\method{y}{vast}(x, ...)

y(x, ...)
}
\arguments{
\item{x}{a \code{vast} object}

\item{...}{ignored}
}
\value{
numeric vector of length \code{nrow(x)}
}
\description{
Cell-centre y coordinates
}
