% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Align the two string and draw the plot of the alignment}
\usage{
align(x, y)
}
\arguments{
\item{x}{A string}

\item{y}{A string}
}
\value{
ggplot of the alignment of x and y
}
\description{
Align the two string and draw the plot of the alignment
}
\examples{
align("ACT", "AGCT")
align("GAGCGT", "GACT")
align("ACTCGCAATATGVTAGGVVAG", "ACTTTATGCTATGCGC")
}
