% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-package.R
\docType{package}
\name{align}
\alias{align-package}
\alias{align}
\title{Data Alignment Tools}
\description{
Methods for the alignment and merging of partially or
inexactly aligned data.
}
\details{
This is a splat packed in-development package.

Package functions:

Two object (x, y, ...) alignment: n_align, cor_align, cow_align.

One object (x, ...) modification: regularize, warp.
}
\references{
Carslaw, D.C., Ropkins, K., Laxen, D., Moorcroft, S., Marner, B. and
Williams, M.L., 2008. Near-field commercial aircraft contribution to
nitrogen oxides by engine, aircraft type, and airline by
individual plume sampling. Environmental science & technology, 42(6),
pp.1871-1876. \doi{https://doi.org/10.1021/es071926a}.

Ropkins, K., Carlsaw, D.C., Goodman, P.S. and Tate, J.E., 2009.
Application of non-linear time-alignment and integration methods to
environmental time series. TrAC Trends in Analytical Chemistry, 28(3),
pp.373-391. \doi{https://doi.org/10.1016/j.trac.2008.11.013}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/karlropkins/align}
  \item \url{https://karlropkins.github.io/align/}
  \item Report bugs at \url{https://github.com/karlropkins/align/issues}
}

}
\author{
\strong{Maintainer}: Karl Ropkins \email{karl.ropkins@gmail.com}

Authors:
\itemize{
  \item Daniel Quiroz Moreno \email{cristianquirozd1997@gmail.com}
}

}
