% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-package.R
\docType{package}
\name{align}
\alias{align}
\title{Data Alignment Tools}
\description{
Methods for the alignment and merging of partially or inexactly
aligned data.
}
\details{
This is a splat packed in-development package.

Package functions:

Two object (x, y, ...) alignment: n_align, cor_align, cow_align.

One object (x, ...) modification: regularize, warp.
}
