% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_align.R
\name{cor_align}
\alias{cor_align}
\title{cor_align}
\usage{
cor_align(x, y = NULL, by = NULL, ...)
}
\arguments{
\item{x}{First \code{vector} or \code{data.frame}, to be aligned
and merged with.}

\item{y}{Second \code{vector} or \code{data.frame}, to align and
merge with \code{x}.}

\item{by}{If \code{x} or \code{y} are \code{data.frame}s, the names
of the columns that \code{cor_align} should align with.}

\item{...}{Other arguments:
\describe{
   \item{\code{min.overlap}}{By default, \code{cor_align} compares
   correlations for lags with \code{x}/\code{y} overlaps of
   2000 rows or 20\% if smaller. Other minimum (row number)
   overlaps can be set using the extra argument
   \code{min.overlap}.}
   \item{\code{output}}{The default \code{cor_align} \code{output}
   is \code{c("plot", "summary", "ans")}. \code{output} options
   include: \code{"ans"}, \code{"plot"}, \code{"summary"} and
   \code{"alignment"}. Multiple \code{output}s are allowed, but
   only the last is captured by return. \code{alignment} is a
   special object class which may be helpful to those looking at
   alignments in more detail.}
}}
}
\value{
By default, \code{cor_align} returns \code{x} and \code{y}
as an aligned \code{data.frame}. It also provides a lag correlation
profile (plot) and alignment report. See above about modifying
outputs.
}
\description{
Correlation based (linear) alignment, row
offsetting second of two \code{vector}s or \code{data.frame}s
based on best fit as estimated by correlation coefficient.
}
\note{
\code{cor_align} is based on \code{cAlign} function in
earlier version of \code{pems.utils}. It works with vectors,
data.frames and other objects that can be converted to data.frame
with as.data.frame(), and uses \code{Rcpp} to run \code{C++} code.
}
\author{
Karl Ropkins
}
