% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow_align.R
\name{cow_align}
\alias{cow_align}
\title{cow_align}
\source{
Based on matlab code at http://www.models.life.ku.dk/dtw_cow.
}
\usage{
cow_align(x, y = NULL, by = NULL, ...)
}
\arguments{
\item{x}{First \code{vector} or a \code{data.frame} containing
\code{vector} to use as reference when COW aligning \code{y} data.}

\item{y}{Second \code{vector} or \code{data.frame} containing
\code{vector} to COW warp align with \code{x}.}

\item{by}{If \code{x} or \code{y} are \code{data.frames}, the
names of data columns to aligned.}

\item{...}{Other arguments, currently ignored.}

\item{seg}{Segment size for warping.}

\item{slack}{Segment size expansion/compression range.}

\item{print.report}{(logical) print segment report.}

\item{r.power}{(logical or numeric) correlation power
(1-4).}
}
\description{
Non-linear alignment using Correlation Optimized
warping (COW). This is \code{cow} but with argument structure like
other ..._align functions.
}
\note{
This function is based on previous matlab (see Source),
but some formal arguments and parameters have been changed to
make function more consistent with other \code{align} functions.
}
\references{
Correlation optimized warping was developed as a
preprocessing method for chromatographic Data:

Niels-Peter Vest Nielsen, Jens Micheal Carstensen
and Jorn Smedegaard,  1998, Aligning of singel and multiple wavelength
chromatographic profiles for chemometric data analysis using
correlation optimised warping. J. Chrom. A 805(1998), 17-35

Giorgio Tomasi, Frans van den Berg and Claus Andersson, 2004,
Correlation optimized warping and dynamic time warping as
preprocessing methods for chromatographic Data, Journal of
Chemometrics 18 (2004), 231-241.
}
\author{
Daniel Quiroz Moreno.
}
