% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_align.R
\name{fit_align}
\alias{fit_align}
\title{fit_align}
\usage{
fit_align(x, y = NULL, by = NULL, ...)
}
\arguments{
\item{x}{A first time-series \code{vector} or a \code{data.frame}
containing time-series to use a reference when warp-fitting
\code{y}.}

\item{y}{A second time-series \code{vector} or a \code{data.frame}
containing time-series to warp-fit with \code{x}.}

\item{by}{If \code{x} or \code{y} are \code{data.frames}, the names
of time-series to warp-fit.}

\item{...}{Other arguments, currently ignored.}

\item{fun}{A \code{function} describing the warp-fitting model to
be applied.}

\item{lower}{The lowest values for \code{fun}, one per parameter.}

\item{upper}{The highest values for \code{fun}, one per parameter.}
}
\description{
Time warp a data-series using a time-offsetting model.
}
\author{
Karl Ropkins
}
