% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{generate_peak}
\alias{generate_peak}
\title{Peak generator.}
\usage{
generate_peak(intensity = 1, length_peak = 100, alpha = 3, xi = -2)
}
\arguments{
\item{intensity}{Maximum intensity of the peak.}

\item{length_peak}{Length of the peak in indexes.}

\item{alpha}{for skewed normal distribution, modifies the first half of the peak}

\item{xi}{for skewed normal distribution, moves the center of the distribution}
}
\value{
A vector with a peak.
}
\description{
Generates a peak with normal distribution.
}
\examples{
generate_peak()
generate_peak(intensity = 500)
generate_peak(length_peak = 150, alpha = 1)
}
\keyword{internal}
