% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_signal.R
\name{generate_peak}
\alias{generate_peak}
\title{Peak generator.}
\usage{
generate_peak(intensity = 1, length_peak = 100, width = 5)
}
\arguments{
\item{intensity}{Maximum intensity of the peak.}

\item{length_peak}{Length of the peak in indexes.}

\item{width}{Width of the peak, this parameter will be passed to 'dnorm()'
and used as sd = length_peak / width.}
}
\value{
A vector with a peak.
}
\description{
Generates a peak with normal distribution.
}
\examples{
generate_peak()
generate_peak(intensity = 500)
generate_peak(length_peak = 150, width = 20)
}
