% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation}
\alias{interpolation}
\title{Interpolation.}
\usage{
interpolation(w, x, return = TRUE)
}
\arguments{
\item{w}{points at which to compute interpolated values.}

\item{x}{signal to be interpolated.}

\item{return}{if TRUE it will return f,s, and g; if FALSE it will return the warped x}
}
\value{
a list with:
f: the interpolated signal,
s: points  of t with  1<= t & t <= length of y
g: gradient of y at points t.
or the warped x
}
\description{
Interpolates a vector.
}
\examples{
w <- c(10:700)
x <- synthetic_signal()
interpolation(w, x)
}
