% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{pow}
\alias{pow}
\title{Penalized Optimized Warping}
\usage{
pow(
  x,
  y,
  lambda1,
  lambda2 = 10^6,
  W = NULL,
  max_it = 1000,
  min_drms = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{x}{Signal to align.}

\item{y}{A reference "signal".}

\item{lambda1}{Regularization parameter for second derivative of warp}

\item{lambda2}{Penalty for negative values of firs derivative (warp)}

\item{W}{A 'diagonal matrix' with weights.}

\item{max_it}{Maximum amount of iterations.}

\item{min_drms}{minimum difference in absolute value of RMS.}

\item{verbose}{by default 'FALSE'to privide information in the console}
}
\value{
A 'list' with:
w: Warping function
sel: part of y that can be fitted
rms: prediction error.
}
\description{
Alignment with POW method.
}
\details{
POW like PTW but without a fix polynomial, it will work
}
\examples{

x <- c(1:10)
y <- c(1:10)
pow(x, y)
}
