% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_reference.R
\name{select_reference}
\alias{select_reference}
\title{Select reference sample}
\usage{
select_reference(data, figure = "correlation")
}
\arguments{
\item{data}{a matrix with all the samples (by row)}

\item{figure}{'variance' or 'correlation' figure of merit to select best reference sample}
}
\value{
index of best reference sample
}
\description{
Select reference sample
}
\examples{
data <- create_synthetic_dataset(n_samples = 10,
                                 n_peaks = 7,
                                 length_out = 100,
                                 mov_peaks = 5)
select_reference(data)
}
