% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{synthetic_chromatogram}
\alias{synthetic_chromatogram}
\title{Synthetic chromatogram.}
\usage{
synthetic_chromatogram(
  n_peaks,
  length_out = 2000,
  intensity = 1000,
  random_intensity = NULL,
  mov_peaks = NULL,
  alpha = 3,
  xi = -2
)
}
\arguments{
\item{n_peaks}{Number of peaks to be added}

\item{length_out}{length of the signal to be created}

\item{intensity}{average intensity of the peaks}

\item{random_intensity}{random maximum value to add or subtract to the
intensity}

\item{mov_peaks}{randomness of the peaks, the smaller it is the most each
peak can move}

\item{alpha}{for skewed normal distribution, modifies the first half of the
peak}

\item{xi}{for skewed normal distribution, moves the center of the
distribution}
}
\value{
a vector that contains a synthetic chromatogram
}
\description{
Creates a synthetic chromatogram.
}
