% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{val}
\alias{val}
\title{Validation}
\usage{
val(X, y, W, fom, iv, lambdas)
}
\arguments{
\item{X}{Matrix with data to be aligned (one sample per row).}

\item{y}{Reference sample.}

\item{W}{A matrix with of weights.}

\item{fom}{figure of merit to use ('rms' or 'cor').}

\item{iv}{indexes of the points used to validate.}

\item{lambdas}{a vector with the penalties to test the POW.}
}
\value{
list with e and i
}
\description{
Validation
}
\keyword{internal}
